refactor(cli): split ServeArgs into cli/serve_args.rs to stay under 500-LoC cap
The MTP gate added --force-speculative (+11 lines), pushing cli.rs from 491
to 502 LoC, over the repo's 500-line-per-file CI cap. Move the ServeArgs clap
struct into a cli/serve_args.rs submodule and re-export it as cli::ServeArgs;
call sites are unchanged. cli.rs is now 21 LoC, serve_args.rs 492.
